[Egg] eggController and buttons

+++ Assets/_games/_common/_scripts/_mustbeimplementedbyhub/SampleWordProvider.cs
@@ -1,13 +1,11 @@
-﻿
-
-namespace EA4S
+﻿namespace EA4S
 {
     /// <summary>
     /// Word provider sample
     /// </summary>
     public class SampleWordProvider : ILivingLetterDataProvider
     {
-        public SampleWordProvider(int difficulty)
+        public SampleWordProvider(float difficulty)
         {
 
         }
